+++ libs/common/src/lib/resources/resource-list/resources-table/resources-table.component.ts
@@ -26,6 +26,7 @@ export class ResourcesTableComponent extends ResourcesTableDirective implements 
   isFiltering = combineLatest([this.resourceListService.filters, this.resourceListService.query]).pipe(
     map(([filters, query]) => filters.length > 0 || query !== ''),
   );
+  hiddenResourcesEnabled = this.resourceListService.hiddenResourcesEnabled;
 
   get initialColumns(): ColumnHeader[] {
     return [
@@ -70,13 +71,19 @@ export class ResourcesTableComponent extends ResourcesTableDirective implements 
 
   private _visibleColumnDef: Observable<ColumnHeader[]> = combineLatest([
     this.isAdminOrContrib,
+    this.hiddenResourcesEnabled,
     this.columnVisibilityUpdate,
   ]).pipe(
-    map(([canEdit]) => {
+    map(([canEdit, hiddenResourcesEnabled]) => {
       const visibleColumns = this.defaultColumns.map(this.getApplySortingMapper()).filter((column) => column.visible);
-      return canEdit
-        ? [...visibleColumns, { id: 'menu', label: 'generic.actions', size: '96px' }]
-        : [...visibleColumns];
+      const extraColumns: ColumnHeader[] = [];
+      if (hiddenResourcesEnabled) {
+        extraColumns.push({ id: 'visibility', label: 'resource.visibility', size: '96px', visible: true });
+      }
+      if (canEdit) {
+        extraColumns.push({ id: 'menu', label: 'generic.actions', size: '96px' });
+      }
+      return [...visibleColumns, ...extraColumns];
     }),
   );
   override headerCells: Observable<HeaderCell[]> = this._visibleColumnDef.pipe(