+++ libs/sdk-core/src/lib/db/kb/kb.models.ts
@@ -39,6 +39,8 @@ export interface IKnowledgeBoxCreation {
   uuid?: string;
   allowed_origins?: string[] | null;
   search_configs?: { [key: string]: any };
+  hidden_resources_enabled?: boolean;
+  hidden_resources_hide_on_creation?: boolean;
 }
 
 export interface IKnowledgeBoxItem extends IKnowledgeBoxCreation {