+++ libs/sdk-core/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix: do not override `show` parameter when calling `ask()`
 - Support `debug` parameter on `/ask` endpoint
 - Add `metadata` and `promptContext` to `Ask.Answer` model
+- Add support for hidden resources
 
 # 1.21.1 (2024-09-12)